// src/api-client/models/ConnectionStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Connected when we can reach the cloud
 * @export
 */
export const ConnectionStatus = {
  Connected: "connected",
  Connecting: "connecting",
  NotReachable: "not_reachable",
  Disconnected: "disconnected",
  BadCredentials: "bad_credentials",
  Null: "null",
} as const;
export type ConnectionStatus =
  (typeof ConnectionStatus)[keyof typeof ConnectionStatus];

export function ConnectionStatusFromJSON(json: any): ConnectionStatus {
  return ConnectionStatusFromJSONTyped(json, false);
}

export function ConnectionStatusFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ConnectionStatus {
  return json as ConnectionStatus;
}

export function ConnectionStatusToJSON(value?: ConnectionStatus | null): any {
  return value as any;
}